Split the seq2seq model into separate files
