+++ src/models/seq2seq/attention.py
@@ -1,0 +1,7 @@
+import tensorflow as tf
+
+
+# TODO
+class Attention(tf.keras.layers.Layer):
+    def __init__(self):
+        pass
