+++ src/models/seq2seq/model.py
@@ -9,199 +9,13 @@ import json
 from collections import Counter
 from itertools import takewhile
 
+from src.models.seq2seq.encoder import Encoder
+from src.models.seq2seq.decoder import Decoder
 from src.metrics.f1_score import F1Score
 from src.preprocessing.tokens import tokenize_method
 
 
-class Encoder(tf.keras.Model):
-    def __init__(
-        self,
-        input_vocab_size,
-        embedding_dims,
-        rnn_units,
-        batch_size, # TODO: can we not pass the batch_size?
-        *args,
-        **kwargs,
-    ):
-        super().__init__(self, args, kwargs)
-
-        self.config = {
-            'input_vocab_size': input_vocab_size,
-            'embedding_dims': embedding_dims,
-            'rnn_units': rnn_units,
-            'batch_size': batch_size,
-        }
-
-        self.embedding = tf.keras.layers.Embedding(
-            input_dim=self.config['input_vocab_size'],
-            output_dim=self.config['embedding_dims'],
-            name='EncoderEmbedding',
-        )
-
-        self.encoder_rnn = tf.keras.layers.LSTM(
-            self.config['rnn_units'],
-            return_sequences=True,
-            return_state=True,
-            name='EncoderLSTM',
-            # default kernel_initializer is 'glorot_uniform',
-            # default recurrent_initializer is 'orthogonal'
-            # default bias_initializer is 'zeros'
-        )
-
-        self.clear_initial_cell_state(batch_size=self.config['batch_size'])
-
-
-    def clear_initial_cell_state(self, batch_size):
-        self.encoder_initial_cell_state = [
-            tf.zeros((batch_size, self.config['rnn_units'])),
-            tf.zeros((batch_size, self.config['rnn_units'])),
-        ]
-
-
-    # annotating with @tf.function leads to None gradients while training
-    # @tf.function(input_signature=[tf.TensorSpec(shape=(None, None))])
-    def call(self, input_batch):
-        output, last_step_hidden_state, last_step_memory_state = self.encoder_rnn(
-            self.embedding(input_batch),
-            initial_state=self.encoder_initial_cell_state
-        )
-
-        return (
-            output, # [batch_size, input sequence length, rnn_units]
-            last_step_hidden_state, # [batch_size, rnn_units]
-            last_step_memory_state # [batch_size, rnn_units]
-        )
-
-    def get_config(self):
-        return self.config
-
-
-class Decoder(tf.keras.Model):
-    def __init__(
-        self,
-        max_output_seq_length,
-        output_vocab_size,
-        embedding_dims,
-        rnn_units,
-        dense_units,
-        batch_size,
-        *args,
-        **kwargs,
-    ):
-        super().__init__(self, args, kwargs)
-
-        self.config = {
-            'max_output_seq_length': max_output_seq_length,
-            'output_vocab_size': output_vocab_size,
-            'embedding_dims': embedding_dims,
-            'rnn_units': rnn_units,
-            'dense_units': dense_units,
-            'batch_size': batch_size,
-        }
-
-        self.embedding = tf.keras.layers.Embedding(
-            input_dim=self.config['output_vocab_size'],
-            output_dim=self.config['embedding_dims'],
-            name='DecoderEmbedding'
-        )
-
-        # TODO: why isn't the activation softmax?
-        self.dense_layer = tf.keras.layers.Dense(
-            self.config['output_vocab_size'],
-            name='DenseOutput'
-        )
-
-        self.decoder_rnn_cell = tf.keras.layers.LSTMCell(
-            self.config['rnn_units'],
-            name='DecoderLSTMCell'
-        )
-
-        # TODO: why prefer Luong over tfa.seq2seq.BahdanauAttention or vice-versa?
-        self.attention_mechanism = tfa.seq2seq.LuongAttention(
-            self.config['dense_units'],
-            memory = None,
-            memory_sequence_length = self.config['batch_size'] * [self.config['max_output_seq_length']]
-        )
-
-        self.rnn_cell = tfa.seq2seq.AttentionWrapper(
-            self.decoder_rnn_cell,
-            self.attention_mechanism,
-            attention_layer_size=self.config['dense_units'],
-        )
-
-        # TODO: isn't this sampler only for training? what if we need to pass a sampler for Beam Search?
-        self.sampler = tfa.seq2seq.sampler.TrainingSampler()
-
-        self.decoder = tfa.seq2seq.BasicDecoder(
-            self.rnn_cell,
-            sampler=self.sampler,
-            output_layer=self.dense_layer,
-        )
-
-        self.setup_memory_and_initial_state() # setup memory with zeros, since we don't have encoder outputs
-
-
-    def build_decoder_initial_state(
-        self,
-        batch_size,
-        encoder_state,
-    ):
-        decoder_initial_state = self.rnn_cell.get_initial_state(
-            batch_size=batch_size,
-            dtype=tf.float32, # TODO: do we need this dtype at all?
-        )
-
-        # TODO: why clone? do we clone the encoder_state? what's going on here?
-        return decoder_initial_state.clone(cell_state=encoder_state)
-
-
-    def setup_memory_and_initial_state(
-        self,
-        encoder_outputs=None,
-        encoder_states=None,
-        batch_size=None
-    ):
-        self.attention_mechanism.setup_memory(
-            encoder_outputs if encoder_outputs is not None else tf.zeros((
-                self.config['batch_size'],
-                200, # TODO: provide input sequence length
-                self.config['rnn_units'],
-            ))
-        )
-
-        self.decoder_initial_state = self.build_decoder_initial_state(
-            # the batch size can be different when in prediction mode
-            batch_size if batch_size is not None else self.config['batch_size'],
-
-            # [last step activations, last memory_state] of encoder is passed as input to decoder Network
-            encoder_state = encoder_states if encoder_states is not None else [
-                tf.zeros((self.config['batch_size'], self.config['rnn_units'])),
-                tf.zeros((self.config['batch_size'], self.config['rnn_units'])),
-            ],
-        )
-
-    # annotating with @tf.function leads to None gradients while training
-    # @tf.function(input_signature=[tf.TensorSpec(shape=(None, None))])
-    def call(self, input_batch):
-        # TODO: document that the memory must be set up with encoder outputs before calling call()
-
-        outputs, hidden_state, cell_state = self.decoder(
-            self.embedding(input_batch),
-            initial_state=self.decoder_initial_state,
-
-            # TODO: don't we know the BATCH_SIZE already inside the decoder? should we?
-            # output sequence length - 1 because of teacher forcing
-            sequence_length=self.config['batch_size'] * [self.config['max_output_seq_length'] - 1]
-        )
-
-        return outputs, hidden_state, cell_state
-
-
-    def get_config(self):
-        return self.config
-
-
-class Seq2SeqAttention(tf.Module):
+class Seq2Seq(tf.Module):
     def __init__(
         self,
         checkpoint_dir,
@@ -375,7 +189,7 @@ class Seq2SeqAttention(tf.Module):
 
         print('Loaded model config: ', config)
 
-        model = Seq2SeqAttention(
+        model = Seq2Seq(
             checkpoint_dir = checkpoint_dir,
             input_vocab_index = input_vocab_index,
             output_vocab_index = output_vocab_index,
