+++ src/server/main.py
@@ -5,7 +5,7 @@ import json
 from flask import Flask
 from flask import request
 
-from src.models.seq2seq_attention import Seq2SeqAttention
+from src.models.seq2seq import Seq2Seq
 
 
 parser = argparse.ArgumentParser(description='Seq2Seq model server')
@@ -38,7 +38,7 @@ def initialize_model(args):
         vocab_path=args.vocab_path
     )
 
-    model = Seq2SeqAttention.restore(
+    model = Seq2Seq.restore(
         args.file_checkpoint_dir,
         input_vocab_index,
         output_vocab_index
